Finally fixed bug tag data appear in quote data for full search result
The bug was caused by the quote absinthe schema where the id from the
the tables (quotes, sources, tags, source types) where represented as
:id in the schema (natural instinct). The problem is that when apollo
at the frontend sees an ID graphQl field, it expects every ID to be
unique. So if I have quote.id 17 and tag.id 17, quote.id will be
replaced by tag.id 17 because apollo assumes there must be only one ID
17. This also the reason React kept complaining about duplicate keys.

The solution is to changed the field from :id to :tid.